+++ packages/nextclaw/src/cli/shared/lib/package-version/index.ts
@@ -1,0 +1,1 @@
+export { readNextclawPackageVersion, resolveNextclawPackageResource } from "./package-version.utils.js";